Update Expo Router + Merging pages
