+++ frontend/src/screens/HomeScreen.tsx
@@ -6,11 +6,13 @@ import {
   TouchableOpacity,
   ScrollView,
   Button,
+  Pressable,
 } from "react-native";
 import { TripCard } from "../components/TripCard";
 import { HomeScreenHeader } from "../components/HomeScreenHeader";
 import { useState, useEffect } from "react";
 import { UserProps, Trip } from "../types";
+import { Link, router } from 'expo-router';
 
 export const HomeScreen: React.FC<UserProps> = ({ user }) => {
   const [ongoingTrips, setOngoingTrips] = useState<Trip[]>([]);
@@ -18,9 +20,11 @@ export const HomeScreen: React.FC<UserProps> = ({ user }) => {
   useEffect(() => {
     const fetchTrips = async () => {
       try {
-        const ongoing = await fetch("http://localhost:3000/trips?ongoing=true");
+        // const ongoing = await fetch("http://localhost:3000/trips?ongoing=true");
+        const ongoing = await fetch("http://10.0.2.2:3000/trips?ongoing=true");
         const upcoming = await fetch(
-          "http://localhost:3000/trips?upcoming=true",
+          // "http://localhost:3000/trips?upcoming=true",
+          "http://10.0.2.2:3000/trips?upcoming=true",
         );
         const ongoingData = await ongoing.json();
         const upcomingData = await upcoming.json();
@@ -48,19 +52,16 @@ export const HomeScreen: React.FC<UserProps> = ({ user }) => {
           <Text style={styles.noTrip}>No ongoing trips</Text>
         )}
       </ScrollView>
-      <View style={styles.buttonContainer}>
-        <TouchableOpacity
-          style={styles.button}
-          onPress={() => {
-            /* navigate to create trip */
-          }}
-        >
-          <Text style={styles.buttonText}>Create a new trip</Text>
-        </TouchableOpacity>
-      </View>
+
+      <Pressable style={styles.buttonContainer}>
+        <Link href="/trips/create" style={styles.button}>
+              <Text style={styles.buttonText}>Create a new trip</Text>
+        </Link>
+      </Pressable>
+
       <View style={styles.upcoming}>
         <Text style={styles.title}>Upcoming Trips</Text>
-        <Button title="See all" />
+        <Button title="See all" onPress={() => {router.replace('/trips/upcoming');}}/>
       </View>
 
       <ScrollView horizontal={true} style={styles.tripScroll}>
@@ -109,6 +110,7 @@ const styles = StyleSheet.create({
     marginTop: 10,
     width: 200,
     alignItems: "center",
+    textAlign: "center",
   },
   buttonText: {
     color: "#FFFFFF",
