+++ frontend/src/screens/UpcomingList.tsx
@@ -6,7 +6,7 @@ import { TripCard } from "../components/TripCard";
 import { LinearGradient } from "expo-linear-gradient";
 import { Trip } from "../types";
 import { AntDesign, MaterialIcons } from '@expo/vector-icons';
-
+import { Stack, Link } from "expo-router";
 
 
 
@@ -19,10 +19,12 @@ const Header = () => (
     <StatusBar backgroundColor="black"/>
     <View style={styles.headerConainner}>
     <View style={styles.headerTitle}>
+      <Link href="/index">
         <TouchableWithoutFeedback onPress={onPressCategory}>
           <AntDesign name="left" size={24} color="blue" />
         </TouchableWithoutFeedback>
-        <Text style={{fontSize: 22, marginLeft: 10}}>Upcoming Trips</Text>
+      </Link>
+      <Text style={{fontSize: 22, marginLeft: 10}}>Upcoming Trips</Text>
     </View>
         
     <View style={styles.userInput}>
@@ -94,7 +96,15 @@ export const ListFilteredCards: React.FC = () => {
 
 
   return(
+    <View>
+        <Stack.Screen
+          options={{
+            title: "",
+            headerShown: true,
+          }}
+        />
       <View style={styles.container}>
+
         <Header/>
         <View style={{flex: 1}}>
           <LinearGradient
@@ -112,6 +122,7 @@ export const ListFilteredCards: React.FC = () => {
             />
           </LinearGradient>
         </View>
+      </View>
       </View>
   );
 };
